make migrations work with multiple schemas move seeding scripts to new model fixes of the app after testing

+++ scripts/purge.script.ts
@@ -1,7 +1,11 @@
 import {truncateSync} from "fs";
+import {Transaction} from "sequelize";
+import {TenantService} from "../src/config/tenant.config";
 import sequelize from "../src/database/sequelize";
+import {TransactionManager} from "../src/database/transaction.manager";
 
 const adminEmail: string = process.env.ADMIN_EMAL || "";
+const tenant: string = process.env.TENANT || "";
 
 if (!adminEmail) {
   console.error("ADMIN_EMAL not found in .env");
@@ -9,6 +13,11 @@ if (!adminEmail) {
 }
 if (process.env.NODE_ENV !== "test") {
   console.error("NODE_ENV must equal 'test'. Add it to .env file.");
+  process.exit(1);
+}
+
+if (!tenant) {
+  console.error("TENANT not found in .env");
   process.exit(1);
 }
 
@@ -28,15 +37,26 @@ export class Purge {
   }
 
   public async purge() {
-    await this.deleteIssues();
-    await this.deleteSprints();
-    await this.deleteUserProjectAccess();
-    await this.deleteProjects();
-    await this.deleteUsers();
+    const schemaName = TenantService.getTenantById(tenant).schemaName;
+
+    const transaction = await TransactionManager.createTenantTransaction(schemaName);
+
+    try {
+      await this.deleteIssues(transaction);
+      await this.deleteSprints(transaction);
+      await this.deleteUserProjectAccess(transaction);
+      await this.deleteProjects(transaction);
+      await this.deleteUsers(transaction);
+
+      await transaction.commit();
+    } catch (error) {
+      await transaction.rollback();
+      throw error;
+    }
 
     try {
       truncateSync("./passwords.txt");
-    } catch (error) {
+    } catch (error: any) {
       if (error.code === "ENOENT") {
         console.log("Passwords file not found");
       } else {
@@ -46,28 +66,30 @@ export class Purge {
     console.log("Passwords file truncated");
   }
 
-  private async deleteIssues() {
-    await sequelize.query("DELETE FROM issue WHERE true");
+  private async deleteIssues(transaction: Transaction) {
+    await sequelize.query("DELETE FROM issue WHERE true", {transaction});
   }
 
-  private async deleteSprints() {
-    await sequelize.query("DELETE FROM sprint WHERE true");
-    await sequelize.query("ALTER SEQUENCE sprint_sprint_id_seq RESTART WITH 1");
+  private async deleteSprints(transaction: Transaction) {
+    await sequelize.query("DELETE FROM sprint WHERE true", {transaction});
+    await sequelize.query("ALTER SEQUENCE sprint_sprint_id_seq RESTART WITH 1", {transaction});
   }
 
-  private async deleteUserProjectAccess() {
-    await sequelize.query("DELETE FROM user_project_access WHERE true");
-    await sequelize.query("ALTER SEQUENCE user_project_access_id_seq RESTART WITH 1");
+  private async deleteUserProjectAccess(transaction: Transaction) {
+    await sequelize.query("DELETE FROM user_project_access WHERE true", {transaction});
+    await sequelize.query("ALTER SEQUENCE user_project_access_id_seq RESTART WITH 1", {transaction});
   }
 
-  private async deleteProjects() {
-    await sequelize.query("DELETE FROM project WHERE true");
-    await sequelize.query("ALTER SEQUENCE project_project_id_seq RESTART WITH 1");
+  private async deleteProjects(transaction: Transaction) {
+    await sequelize.query("DELETE FROM project WHERE true", {transaction});
+    await sequelize.query("ALTER SEQUENCE project_project_id_seq RESTART WITH 1", {transaction});
   }
 
-  private async deleteUsers() {
-    await sequelize.query(`DELETE FROM admin_user WHERE user_id != (SELECT user_id FROM zilla_user WHERE email = '${adminEmail}')`);
-    await sequelize.query(`DELETE FROM zilla_user WHERE email != '${adminEmail}'`);
+  private async deleteUsers(transaction: Transaction) {
+    await sequelize.query(`DELETE FROM admin_user WHERE user_id != (SELECT user_id FROM zilla_user WHERE email = '${adminEmail}')`, {
+      transaction,
+    });
+    await sequelize.query(`DELETE FROM zilla_user WHERE email != '${adminEmail}'`, {transaction});
   }
 }
